Do not show unsaved changes popup when policy form is not dirty
