+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add request policy CRU v1. Fixes UIREQ-181.
 * Add permission set for staff slips. Fixes UICIRC-194.
 * Update circulation OKAPI interface to v7.0. Part of UICHKOUT-492.
+* Do not show unsaved changes popup when policy form is not dirty. Fixes UICIRC-104.
 
 ## 1.5.0 (https://github.com/folio-org/ui-circulation/tree/v1.5.0) (2019-01-25)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v1.4.0...v1.5.0)
