+++ src/constants.js
@@ -33,6 +33,7 @@ export const CURRENT_DUE_DATE_TIME = 'CURRENT_DUE_DATE_TIME';
 export const END_OF_THE_CURRENT_SERVICE_POINT_HOURS = 'END_OF_THE_CURRENT_SERVICE_POINT_HOURS';
 export const BEGINNING_OF_THE_NEXT_OPEN_SERVICE_POINT_HOURS = 'BEGINNING_OF_THE_NEXT_OPEN_SERVICE_POINT_HOURS';
 
+export const KEEP_CURRENT_DATE = 'KEEP_CURRENT_DATE';
 export const CURRENT_DUE_DATE = 'CURRENT_DUE_DATE';
 export const END_OF_THE_PREVIOUS_OPEN_DAY = 'END_OF_THE_PREVIOUS_OPEN_DAY';
 export const END_OF_THE_NEXT_OPEN_DAY = 'END_OF_THE_NEXT_OPEN_DAY';
