+++ src/settings/LoanPolicy/LoanPolicyForm.js
@@ -196,5 +196,5 @@ const connectedLoanPolicyForm = connect(mapStateToProps)(LoanPolicyForm);
 export default stripesForm({
   form: 'loanPolicyForm',
   navigationCheck: true,
-  enableReinitialize: false,
+  enableReinitialize: true,
 })(connectedLoanPolicyForm);
