+++ src/settings/LoanPolicy/components/EditSections/LoansSection/LoansSection.js
@@ -2,10 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'redux-form';
-import {
-  get,
-  some,
-} from 'lodash';
 
 import {
   Checkbox,
@@ -21,8 +17,6 @@ import {
   intervalPeriods,
   shortTermLoansOptions,
   longTermLoansOptions,
-  CURRENT_DUE_DATE_TIME,
-  CURRENT_DUE_DATE,
 } from '../../../../../constants';
 
 class LoansSection extends React.Component {
@@ -30,38 +24,6 @@ class LoansSection extends React.Component {
     policy: PropTypes.object.isRequired,
     schedules: PropTypes.arrayOf(PropTypes.node).isRequired,
     change: PropTypes.func.isRequired,
-  };
-
-  componentDidUpdate() {
-    this.setDueDateManagementSelectedId();
-  }
-
-  isValidItemSelected = (options, selectedId) => {
-    return some(options, ({ id }) => id === selectedId);
-  };
-
-  setDueDateManagementSelectedId = () => {
-    const {
-      policy,
-      change,
-    } = this.props;
-
-    const pathToField = 'loansPolicy.closedLibraryDueDateManagementId';
-    const isShortTermLoan = policy.isShortTermLoan();
-    const selectedId = get(policy, pathToField);
-
-    const isValidShortTermLoanValue = this.isValidItemSelected(shortTermLoansOptions, selectedId);
-    const isValidLongTermLoanValue = this.isValidItemSelected(longTermLoansOptions, selectedId);
-
-    if (isShortTermLoan && !isValidShortTermLoanValue) {
-      /* Set default value for short term loan if long term loan item was selected */
-      change(pathToField, CURRENT_DUE_DATE_TIME);
-    }
-
-    if (!isShortTermLoan && !isValidLongTermLoanValue) {
-      /* Set default value for long term loan if short term loan item was selected */
-      change(pathToField, CURRENT_DUE_DATE);
-    }
   };
 
   render() {
