+++ src/settings/Models/LoanPolicy/LoansPolicy.js
@@ -1,8 +1,16 @@
+import { some } from 'lodash';
 import { Period } from '../common';
 import {
   defaultLoanPolicy,
   helpers,
 } from './utils';
+import {
+  intervalIdsMap,
+  shortTermLoansOptions,
+  longTermLoansOptions,
+  CURRENT_DUE_DATE_TIME,
+  CURRENT_DUE_DATE,
+} from '../../../constants';
 
 export default class LoansPolicy {
   constructor(policy = {}) {
@@ -10,8 +18,8 @@ export default class LoansPolicy {
     this.closedLibraryDueDateManagementId = policy.closedLibraryDueDateManagementId;
     this.fixedDueDateScheduleId = policy.fixedDueDateScheduleId;
     this.period = new Period(policy.period);
-    this.gracePeriod = new Period(policy.gracePeriod);
-    this.openingTimeOffset = new Period(policy.openingTimeOffset);
+    this.gracePeriod = new Period(policy.gracePeriod || { intervalId: intervalIdsMap.HOURS });
+    this.openingTimeOffset = new Period(policy.openingTimeOffset || { intervalId: intervalIdsMap.HOURS });
   }
 
   get defaultsSelected() {
@@ -21,4 +29,23 @@ export default class LoansPolicy {
   get additionalFieldsSelected() {
     return helpers.additionalFieldsSelected(this, defaultLoanPolicy.loansPolicy);
   }
+
+  // set correct closedLibraryDueDateManagementId
+  setDueDateManagementId(isShortTermLoan) {
+    const selectedId = this.closedLibraryDueDateManagementId;
+    const isValidShortTermLoanValue = this.isValidItemSelected(shortTermLoansOptions, selectedId);
+    const isValidLongTermLoanValue = this.isValidItemSelected(longTermLoansOptions, selectedId);
+
+    if (isShortTermLoan && !isValidShortTermLoanValue) {
+      this.closedLibraryDueDateManagementId = CURRENT_DUE_DATE_TIME;
+    }
+
+    if (!isShortTermLoan && !isValidLongTermLoanValue) {
+      this.closedLibraryDueDateManagementId = CURRENT_DUE_DATE;
+    }
+  }
+
+  isValidItemSelected = (options, selectedId) => {
+    return some(options, ({ id }) => id === selectedId);
+  };
 }
