+++ src/settings/Models/LoanPolicy/RenewalsPolicy.js
@@ -3,12 +3,15 @@ import {
   defaultLoanPolicy,
   helpers,
 } from './utils';
+import {
+  intervalIdsMap,
+} from '../../../constants';
 
 export default class RenewalsPolicy {
   constructor(policy = {}) {
     this.unlimited = policy.unlimited || false;
     this.numberAllowed = policy.numberAllowed;
-    this.period = new Period(policy.period);
+    this.period = new Period(policy.period || { intervalId: intervalIdsMap.DAYS });
     this.renewFromId = policy.renewFromId;
     this.differentPeriod = policy.differentPeriod || false;
     this.alternateFixedDueDateScheduleId = policy.alternateFixedDueDateScheduleId;
